--- src/test/java/desafio/grupo2/rumbo/testcreation/testcases/CP006_Hoteles.java
@@ -1,31 +1,0 @@
-package desafio.grupo2.rumbo.testcreation.testcases;
-
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsHomePage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsHotelesPage;
-import framework.engine.selenium.DriverFactory;
-import framework.engine.selenium.SeleniumTestBase;
-import io.qameta.allure.Description;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-class CP006_Hoteles extends SeleniumTestBase {
-
-    RumboEsHomePage rumboEsHomePage;
-
-    RumboEsHotelesPage rumboEsHotelesPage;
-
-    @Test
-    @Description("Test Caso CP006_Hoteles")
-    void CP006_Hoteles() throws InterruptedException {
-        rumboEsHomePage = new RumboEsHomePage(DriverFactory.getDriver());
-        rumboEsHomePage.despegarARumbos();
-        rumboEsHomePage.aceptarCookies();
-        rumboEsHomePage.irAHoteles();
-
-        Assertions.assertEquals("Hoteles | Ofertas de Hoteles baratos | Rumbo", rumboEsHomePage.getUrlTitle());
-
-        rumboEsHotelesPage = new RumboEsHotelesPage(DriverFactory.getDriver());
-        rumboEsHotelesPage.ingresarDestino("Chicago");
-    }
-}
-
