+++ src/test/java/desafio/grupo2/rumbo/testcreation/testcases/testHoteles/CP005_Hoteles.java
@@ -1,21 +1,22 @@
-package desafio.grupo2.rumbo.testcreation.testcases;
+package desafio.grupo2.rumbo.testcreation.testcases.testHoteles;
 
 import desafio.grupo2.rumbo.testcreation.pages.RumboEsHomePage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsHotelesBusquedaPage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsHotelesPage;
+import desafio.grupo2.rumbo.testcreation.pages.Hoteles.RumboEsHotelesPage;
 import framework.engine.selenium.DriverFactory;
 import framework.engine.selenium.SeleniumTestBase;
 import io.qameta.allure.Description;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-class CP004_Hoteles extends SeleniumTestBase {
+class CP005_Hoteles extends SeleniumTestBase {
+
     RumboEsHomePage rumboEsHomePage;
+
     RumboEsHotelesPage rumboEsHotelesPage;
-    RumboEsHotelesBusquedaPage rumboEsHotelesBusquedaPage;
+
     @Test
-    @Description("Test Caso CP004_Hoteles")
-    void CP004_Hoteles() throws InterruptedException {
+    @Description("Test Caso CP005_Hoteles")
+    void CP005_Hoteles() throws InterruptedException {
         rumboEsHomePage = new RumboEsHomePage(DriverFactory.getDriver());
         rumboEsHomePage.despegarARumbos();
         rumboEsHomePage.aceptarCookies();
@@ -25,15 +26,6 @@ class CP004_Hoteles extends SeleniumTestBase {
 
         rumboEsHotelesPage = new RumboEsHotelesPage(DriverFactory.getDriver());
         rumboEsHotelesPage.ingresarDestino("Chicago");
-        rumboEsHotelesPage.FechaInicioVuelta();
-        rumboEsHotelesPage.pasajeros();
-        rumboEsHotelesPage.buscar();
-
-        Assertions.assertEquals("Inspírate para tus viajes en 2022/2023 | lastminute.com", rumboEsHomePage.getUrlTitle());
-
-        rumboEsHotelesBusquedaPage = new RumboEsHotelesBusquedaPage(DriverFactory.getDriver());
-        rumboEsHotelesBusquedaPage.seleccionarHotel2();
-
     }
 }
