+++ src/test/java/desafio/grupo2/rumbo/testcreation/testcases/testHoteles/CP006_Hoteles.java
@@ -1,22 +1,22 @@
-package desafio.grupo2.rumbo.testcreation.testcases;
+package desafio.grupo2.rumbo.testcreation.testcases.testHoteles;
 
 import desafio.grupo2.rumbo.testcreation.pages.RumboEsHomePage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsHotelesPage;
+import desafio.grupo2.rumbo.testcreation.pages.Hoteles.RumboEsHotelesPage;
 import framework.engine.selenium.DriverFactory;
 import framework.engine.selenium.SeleniumTestBase;
 import io.qameta.allure.Description;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-class CP005_Hoteles extends SeleniumTestBase {
+class CP006_Hoteles extends SeleniumTestBase {
 
     RumboEsHomePage rumboEsHomePage;
 
     RumboEsHotelesPage rumboEsHotelesPage;
 
     @Test
-    @Description("Test Caso CP005_Hoteles")
-    void CP005_Hoteles() throws InterruptedException {
+    @Description("Test Caso CP006_Hoteles")
+    void CP006_Hoteles() throws InterruptedException {
         rumboEsHomePage = new RumboEsHomePage(DriverFactory.getDriver());
         rumboEsHomePage.despegarARumbos();
         rumboEsHomePage.aceptarCookies();
