+++ src/test/java/desafio/grupo2/rumbo/testcreation/testcases/testTrenes/CP001_Trenes.java
@@ -1,39 +1,37 @@
-package desafio.grupo2.rumbo.testcreation.testcases;
+package desafio.grupo2.rumbo.testcreation.testcases.testTrenes;
 
 import desafio.grupo2.rumbo.testcreation.pages.RumboEsHomePage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsTrenesBusquedaPage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsTrenesPage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsTrenesSecurePage;
+import desafio.grupo2.rumbo.testcreation.pages.Trenes.RumboEsTrenesBusquedaPage;
+import desafio.grupo2.rumbo.testcreation.pages.Trenes.RumboEsTrenesPage;
 import framework.engine.selenium.DriverFactory;
 import framework.engine.selenium.SeleniumTestBase;
 import io.qameta.allure.Description;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+
 import java.util.stream.Stream;
 
-public class CP002_Trenes extends SeleniumTestBase {
+public class CP001_Trenes extends SeleniumTestBase {
 
     RumboEsHomePage rumboEsHomePage;
     RumboEsTrenesPage rumboEsTrenesPage;
 
     RumboEsTrenesBusquedaPage rumboEsTrenesBusquedaPage;
 
-    RumboEsTrenesSecurePage rumboEsTrenesSecurePage;
-
     @ParameterizedTest
     @MethodSource
-    @Description("Realizar el test CP002 del RF02")
-    void CP002_MenorDuracion(String Origen,String Destino, String esperado) throws InterruptedException {
-    rumboEsHomePage = new RumboEsHomePage(DriverFactory.getDriver());
+    @Description("Realizar el test CP001 del RF02")
+    void CP001_BusquedadeTren(String Origen,String Destino, String esperado) throws InterruptedException {
+
+        rumboEsHomePage = new RumboEsHomePage(DriverFactory.getDriver());
         rumboEsHomePage.despegarARumbos();
         rumboEsHomePage.aceptarCookies();
         rumboEsHomePage.irATrenes();
 
 
-    rumboEsTrenesPage = new RumboEsTrenesPage(DriverFactory.getDriver());
+        rumboEsTrenesPage = new RumboEsTrenesPage(DriverFactory.getDriver());
         rumboEsTrenesPage.seleccionarIda();
         rumboEsTrenesPage.ingresarOrigen(Origen);
         rumboEsTrenesPage.ingresarDestino(Destino);
@@ -41,20 +39,13 @@ public class CP002_Trenes extends SeleniumTestBase {
         rumboEsTrenesPage.establecerBusqueda();
 
         rumboEsTrenesBusquedaPage = new RumboEsTrenesBusquedaPage(DriverFactory.getDriver());
-        rumboEsTrenesBusquedaPage.duracionmenor();
-
-
-
-        rumboEsTrenesBusquedaPage = new RumboEsTrenesBusquedaPage(DriverFactory.getDriver());
         String Resultado = rumboEsTrenesBusquedaPage.obtenerRutaOri() + " - " + rumboEsTrenesBusquedaPage.obtenerRutaDes();
 
         Assertions.assertEquals(esperado,Resultado);
 
 
-
-
-}
-    static Stream<Arguments> CP002_MenorDuracion(){
+    }
+    static Stream<Arguments>CP001_BusquedadeTren(){
         return Stream.of(
 
                 Arguments.arguments("Madrid","Barcelona", "XOC - YJB")
