+++ src/test/java/desafio/grupo2/rumbo/testcreation/testcases/testTrenes/CP002_Trenes.java
@@ -1,9 +1,9 @@
-package desafio.grupo2.rumbo.testcreation.testcases;
+package desafio.grupo2.rumbo.testcreation.testcases.testTrenes;
 
 import desafio.grupo2.rumbo.testcreation.pages.RumboEsHomePage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsTrenesBusquedaPage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsTrenesPage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsTrenesSecurePage;
+import desafio.grupo2.rumbo.testcreation.pages.Trenes.RumboEsTrenesBusquedaPage;
+import desafio.grupo2.rumbo.testcreation.pages.Trenes.RumboEsTrenesPage;
+import desafio.grupo2.rumbo.testcreation.pages.Trenes.RumboEsTrenesSecurePage;
 import framework.engine.selenium.DriverFactory;
 import framework.engine.selenium.SeleniumTestBase;
 import io.qameta.allure.Description;
@@ -11,10 +11,10 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-
 import java.util.stream.Stream;
 
-public class CP003_Trenes extends SeleniumTestBase {
+public class CP002_Trenes extends SeleniumTestBase {
+
     RumboEsHomePage rumboEsHomePage;
     RumboEsTrenesPage rumboEsTrenesPage;
 
@@ -22,18 +22,17 @@ public class CP003_Trenes extends SeleniumTestBase {
 
     RumboEsTrenesSecurePage rumboEsTrenesSecurePage;
 
-
     @ParameterizedTest
     @MethodSource
-    @Description("Realizar el test CP003 del RF02")
-    void CP003_DatosIncompletos(String Origen, String Destino, String esperado) throws InterruptedException {
-        rumboEsHomePage = new RumboEsHomePage(DriverFactory.getDriver());
+    @Description("Realizar el test CP002 del RF02")
+    void CP002_MenorDuracion(String Origen,String Destino, String esperado) throws InterruptedException {
+    rumboEsHomePage = new RumboEsHomePage(DriverFactory.getDriver());
         rumboEsHomePage.despegarARumbos();
         rumboEsHomePage.aceptarCookies();
         rumboEsHomePage.irATrenes();
 
 
-        rumboEsTrenesPage = new RumboEsTrenesPage(DriverFactory.getDriver());
+    rumboEsTrenesPage = new RumboEsTrenesPage(DriverFactory.getDriver());
         rumboEsTrenesPage.seleccionarIda();
         rumboEsTrenesPage.ingresarOrigen(Origen);
         rumboEsTrenesPage.ingresarDestino(Destino);
@@ -41,21 +40,23 @@ public class CP003_Trenes extends SeleniumTestBase {
         rumboEsTrenesPage.establecerBusqueda();
 
         rumboEsTrenesBusquedaPage = new RumboEsTrenesBusquedaPage(DriverFactory.getDriver());
-        rumboEsTrenesBusquedaPage.seleccionarViaje();
-
-        rumboEsTrenesSecurePage = new RumboEsTrenesSecurePage(DriverFactory.getDriver());
-        rumboEsTrenesSecurePage.pasarSiguiente();
-        String resultado = rumboEsTrenesSecurePage.sinDatos();
-
-        Assertions.assertEquals(esperado, resultado);
-    }
+        rumboEsTrenesBusquedaPage.duracionmenor();
 
 
-    static Stream<Arguments> CP003_DatosIncompletos(){
+
+        rumboEsTrenesBusquedaPage = new RumboEsTrenesBusquedaPage(DriverFactory.getDriver());
+        String Resultado = rumboEsTrenesBusquedaPage.obtenerRutaOri() + " - " + rumboEsTrenesBusquedaPage.obtenerRutaDes();
+
+        Assertions.assertEquals(esperado,Resultado);
+
+
+
+
+}
+    static Stream<Arguments> CP002_MenorDuracion(){
         return Stream.of(
 
-                Arguments.arguments("Madrid","Barcelona", "Introduce el nombre")
+                Arguments.arguments("Madrid","Barcelona", "XOC - YJB")
         );
     }
 }
-
