+++ src/test/java/desafio/grupo2/rumbo/testcreation/testcases/testTrenes/CP003_Trenes.java
@@ -1,31 +1,32 @@
-package desafio.grupo2.rumbo.testcreation.testcases;
+package desafio.grupo2.rumbo.testcreation.testcases.testTrenes;
 
 import desafio.grupo2.rumbo.testcreation.pages.RumboEsHomePage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsTrenesBusquedaPage;
-import desafio.grupo2.rumbo.testcreation.pages.RumboEsTrenesPage;
+import desafio.grupo2.rumbo.testcreation.pages.Trenes.RumboEsTrenesBusquedaPage;
+import desafio.grupo2.rumbo.testcreation.pages.Trenes.RumboEsTrenesPage;
+import desafio.grupo2.rumbo.testcreation.pages.Trenes.RumboEsTrenesSecurePage;
 import framework.engine.selenium.DriverFactory;
 import framework.engine.selenium.SeleniumTestBase;
 import io.qameta.allure.Description;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
 
-public class CP001_Trenes extends SeleniumTestBase {
-
+public class CP003_Trenes extends SeleniumTestBase {
     RumboEsHomePage rumboEsHomePage;
     RumboEsTrenesPage rumboEsTrenesPage;
 
     RumboEsTrenesBusquedaPage rumboEsTrenesBusquedaPage;
 
+    RumboEsTrenesSecurePage rumboEsTrenesSecurePage;
+
+
     @ParameterizedTest
     @MethodSource
-    @Description("Realizar el test CP001 del RF02")
-    void CP001_BusquedadeTren(String Origen,String Destino, String esperado) throws InterruptedException {
-
+    @Description("Realizar el test CP003 del RF02")
+    void CP003_DatosIncompletos(String Origen, String Destino, String esperado) throws InterruptedException {
         rumboEsHomePage = new RumboEsHomePage(DriverFactory.getDriver());
         rumboEsHomePage.despegarARumbos();
         rumboEsHomePage.aceptarCookies();
@@ -40,16 +41,21 @@ public class CP001_Trenes extends SeleniumTestBase {
         rumboEsTrenesPage.establecerBusqueda();
 
         rumboEsTrenesBusquedaPage = new RumboEsTrenesBusquedaPage(DriverFactory.getDriver());
-        String Resultado = rumboEsTrenesBusquedaPage.obtenerRutaOri() + " - " + rumboEsTrenesBusquedaPage.obtenerRutaDes();
+        rumboEsTrenesBusquedaPage.seleccionarViaje();
 
-        Assertions.assertEquals(esperado,Resultado);
+        rumboEsTrenesSecurePage = new RumboEsTrenesSecurePage(DriverFactory.getDriver());
+        rumboEsTrenesSecurePage.pasarSiguiente();
+        String resultado = rumboEsTrenesSecurePage.sinDatos();
 
-
+        Assertions.assertEquals(esperado, resultado);
     }
-    static Stream<Arguments>CP001_BusquedadeTren(){
+
+
+    static Stream<Arguments> CP003_DatosIncompletos(){
         return Stream.of(
 
-                Arguments.arguments("Madrid","Barcelona", "XOC - YJB")
+                Arguments.arguments("Madrid","Barcelona", "Introduce el nombre")
         );
     }
 }
+
